requirements.txt
-i https://pypi.org/simple
asgiref==3.10.0; python_version >= '3.9'
cachetools==6.2.1; python_version >= '3.9'
certifi==2025.10.5; python_version >= '3.7'
cffi==2.0.0; python_version >= '3.9'
charset-normalizer==3.4.4; python_version >= '3.7'
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
cryptography==46.0.3; python_version >= '3.8' and python_full_version not in '3.9.0, 3.9.1'
dj-rest-auth[with_social]==7.0.1; python_version >= '3.8'
django==5.2.7; python_version >= '3.10'
django-allauth[socialaccount]==65.13.0; python_version >= '3.8'
django-cors-headers==4.9.0; python_version >= '3.9'
django-rest-passwordreset==1.5.0
djangorestframework==3.16.1; python_version >= '3.9'
djangorestframework-simplejwt==5.5.1; python_version >= '3.9'
google-auth==2.42.1; python_version >= '3.7'
idna==3.11; python_version >= '3.8'
markdown-it-py==4.0.0; python_version >= '3.10'
mdurl==0.1.2; python_version >= '3.7'
oauthlib==3.3.1; python_version >= '3.8'
psycopg==3.2.12; python_version >= '3.8'
psycopg-binary==3.2.12; python_version >= '3.8'
psycopg2-binary==2.9.11; python_version >= '3.9'
pyasn1==0.6.1; python_version >= '3.8'
pyasn1-modules==0.4.2; python_version >= '3.8'
pycparser==2.23; python_version >= '3.8'
pygments==2.19.2; python_version >= '3.8'
pyjwt[crypto]==2.10.1; python_version >= '3.9'
python-decouple==3.8
requests==2.32.5; python_version >= '3.9'
rich==14.2.0; python_full_version >= '3.8.0'
rsa==4.9.1; python_version >= '3.6' and python_version < '4'
sentry-sdk[django]==2.42.1; python_version >= '3.6'
sqlparse==0.5.3; python_version >= '3.8'
typing-extensions==4.15.0; python_version >= '3.9'
tzdata==2025.2; python_version >= '2'
urllib3==2.5.0; python_version >= '3.9'
